rename python file and return json based on stage rather than bool
